03042023_licenta/flaskapp/requirements.txt
click==8.1.3
Flask==2.2.3
gunicorn==20.1.0
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
Werkzeug==2.2.3
zipp==3.15.0
